refactor camera imaging in prep for fast z scan
