+++ acq4/util/imaging/sequencerTemplate.ui
